+++ index.html
@@ -41,7 +41,7 @@
     }
   </style>
 
-  <script type="module" src="/lib/elliott-main-page.js"></script>
+  <script type="module" src="/lib/components/elliott-main-page.js"></script>
 </head>
 <body>
   <noscript>
