+++ src/components/elliott-header-bar.ts
